⬆️ Upgrade dependencies and clean some code
