+++ package.json
@@ -49,11 +49,6 @@
       "/bowling-kata"
     ]
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
       "prettier-eslint --write",
@@ -64,26 +59,24 @@
   "dependencies": {
     "cac": "^6.4.2",
     "chalk": "^2.4.2",
-    "inquirer": "^6.2.2",
+    "inquirer": "^8.2.2",
     "shelljs": "^0.8.3"
   },
   "devDependencies": {
     "@types/inquirer": "^0.0.43",
-    "@types/jest": "^24.0.6",
-    "@types/jest-diff": "^20.0.1",
+    "@types/jest": "^27.4.1",
     "@types/node": "^11.9.4",
     "@types/shelljs": "^0.8.3",
-    "@typescript-eslint/eslint-plugin": "^1.4.2",
-    "@typescript-eslint/parser": "^1.4.2",
-    "eslint": "^5.15.1",
+    "@typescript-eslint/eslint-plugin": "^5.16.0",
+    "@typescript-eslint/parser": "^5.16.0",
+    "eslint": "^8.12.0",
     "eslint-config-prettier": "^4.1.0",
-    "husky": "^1.3.1",
-    "jest": "^24.1.0",
-    "lint-staged": "^8.1.5",
+    "jest": "^27.5.1",
+    "lint-staged": "^12.3.7",
     "prettier": "^1.16.4",
-    "prettier-eslint-cli": "^5.0.0",
-    "ts-jest": "^23.10.5",
-    "ts-node": "^8.0.2",
+    "prettier-eslint-cli": "^4.2.1",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
     "typescript": "^3.3.3"
   },
   "repository": {
