+++ src/lib/katas/fizz-buzz.ts
@@ -1,0 +1,35 @@
+export const FizzBuzz = {
+  title: 'fizz-buzz-kata',
+  description: `
+Write a program that prints the numbers from 1 to 100, but if a number is a multiple of:
+    * 3, print \`fizz\` instead of 3
+    * 5, print \`buzz\` instead of 5
+    * 3 and 5, print \`fizzbuzz\`
+
+Sample output:
+
+\`\`\`
+1
+2
+Fizz
+4
+Buzz
+Fizz
+7
+8
+Fizz
+Buzz
+11
+Fizz
+13
+14
+FizzBuzz
+16
+17
+Fizz
+19
+Buzz
+... etc up to 100
+\`\`\`
+    `,
+}
