+++ src/lib/katas/tic-tac-toe.ts
@@ -1,0 +1,15 @@
+export const TicTacToe = {
+  title: 'tic-tac-toe-kata',
+  description: `
+## Rules
+
+X always goes first.
+
+Players cannot play on a played position.
+
+Players alternate placing X's and O's on the board until either:
+
+  * One player has three in a row, horizontally, vertically or diagonally -> win
+  * All nine squares are filled -> draw
+    `,
+}
