+++ src/lib/katas/xmas-tree.ts
@@ -1,0 +1,20 @@
+export const XmasTree = {
+  title: 'christmas-tree',
+  description: `
+Create a function \`christmasTree(height)\` that returns a christmas tree of the correct height.
+christmasTree(5) should return:
+
+\`\`\`
+          *
+         ***
+        *****
+       *******
+      *********
+\`\`\`
+
+Height passed is always an integer between 0 and 100.
+
+Use \`\n\` for newlines between each line.
+Pad with spaces so each line is the same length. The last line having only stars, no spaces.
+    `,
+}
