+++ templates/javascript/package.json
@@ -1,5 +1,6 @@
 {
   "name": "app_title",
+  "version": "1.0.0",
   "description": "Kata bootstrapped with kataclism",
   "main": "index.js",
   "scripts": {
@@ -24,8 +25,8 @@
     "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/preset-env": "^7.1.6",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^23.6.0",
-    "jest": "^23.6.0",
+    "babel-jest": "^27.5.1",
+    "jest": "^27.5.1",
     "prettier": "^1.15.2"
   }
 }
