+++ templates/typescript/package.json
@@ -1,5 +1,6 @@
 {
   "name": "app_title",
+  "version": "1.0.0",
   "description": "Kata bootstrapped with kataclism",
   "main": "index.js",
   "scripts": {
@@ -26,11 +27,11 @@
     "singleQuote": true
   },
   "devDependencies": {
-    "@types/jest": "^24.0.3",
+    "typescript": "^3.3.3",
     "@types/typescript": "^2.0.0",
-    "jest": "^24.1.0",
-    "ts-jest": "^23.10.5",
-    "ts-node": "^8.0.2",
-    "typescript": "^3.3.3"
+    "@types/jest": "^27.4.1",
+    "jest": "^27.5.1",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0"
   }
 }
